requirements.txt
langchain_community==0.4
langchain_huggingface==1.0.0
sentence_transformers==5.1.1
langchain_groq==1.0.0
unstructured==0.18.15
chromadb==1.2.1
